nyayasathi/requirements.txt
Django==5.2.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
requests==2.31.0
groq==0.4.2
langid==1.1.6
deep-translator==1.11.4
python-docx==1.1.0
PyPDF2==3.0.1
whisper==1.1.10
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
dj-database-url==2.1.0 